client: export the InitStartTLS function
If the initStartTLS function remains unexported, there is no way to get
hold of a *Client after NewClient is called, and before initStartTLS is
called.

This is needed if the Hello() method needs to be called manually to set
a custom host name instead of the default "localhost".

In my specific case, the smtp-relay.gmail.com server rejects my request
if I don't send a host name different of "localhost" in the EHLO command.